@@ -16,9 +16,11 @@
 	<spring.version>3.2.0.RELEASE</spring.version>
 	<log4j.version>1.2.14</log4j.version>
 	<mysql.version>5.1.2</mysql.version>
-	<junit.version>4.8.2</junit.version>
-	<maven.compiler.source>1.6</maven.compiler.source>                                        
-    <maven.compiler.target>1.6</maven.compiler.target>  
+	<java.version>1.8</java.version>
+	<junit-jupiter.version>5.5.2</junit-jupiter.version>
+	<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+	<maven.compiler.source>1.8</maven.compiler.source>                                        
+    <maven.compiler.target>1.8</maven.compiler.target>  
   </properties>
   
   <reporting>
@@ -190,12 +192,19 @@
 		<scope>test</scope>
 		<type>jar</type>
 	</dependency>
+	 <!-- junit 5 -->
 	<dependency>
-		<groupId>junit</groupId>
-		<artifactId>junit</artifactId>
-		<version>${junit.version}</version>
+		<groupId>org.junit.jupiter</groupId>
+		<artifactId>junit-jupiter-engine</artifactId>
+		<version>${junit-jupiter.version}</version>
 		<scope>test</scope>
-		<type>jar</type>
+	</dependency>
+	<!-- mockito -->
+	<dependency>
+		<groupId>org.mockito</groupId>
+		<artifactId>mockito-junit-jupiter</artifactId>
+		<version>3.2.4</version>
+		<scope>test</scope>
 	</dependency>
 	<dependency>
 		<groupId>javax.persistence</groupId>
