@@ -3,7 +3,7 @@
  */
 package com.juxtapose.example.ch07.jms;
 
-import junit.framework.Assert;
+// import junit.framework.Assert;
 
 import org.springframework.batch.core.ExitStatus;
 import org.springframework.batch.core.StepExecution;
@@ -13,8 +13,7 @@ import org.springframework.jms.core.JmsTemplate;
 import com.juxtapose.example.ch07.CreditBill;
 
 /**
- * @author bruce.liu(mailto:jxta.liu@gmail.com)
- * 2013-9-21下午08:51:51
+ * @author taegepn-park23(justerror23@gmail.com)
  * @param <S>
  */
 public class JMSDetectItemWriteListener implements StepExecutionListener{
@@ -33,7 +32,8 @@ public class JMSDetectItemWriteListener implements StepExecutionListener{
 			System.out.println("Receive from jms queue:"+result);
 			obj = jmsTemplate.receiveAndConvert();
 		}
-		Assert.assertEquals(stepExecution.getWriteCount(), writeCount);
+		
+		// Assert.assertEquals(stepExecution.getWriteCount(), writeCount);
 		return stepExecution.getExitStatus();
 	}
 
